refactor : boardlist 삭제 및 오류 해결
